Add files via upload
Note author: Yifan Wang.

CART model for citation prediction.
"data.txt" is the input file. Each line contains the information of a paper. The amount of the numbers in each line should be equal. The last number is which you want to predict, and the numbers ahead are the features. The code divide the input data into a training set and a test set, according to a constant ratio. "out.txt" is the output file. One line contains a predicted number and an actual number of a case in the test set.